Rename tx asset files after classname
