+++ tutorials/nft/blockchain-app/nft-module/transactions/create_nft_asset.js
