+++ tutorials/nft/blockchain-app/nft-module/transactions/purchase_nft_asset.js
