+++ tutorials/nft/blockchain-app/nft-module/transactions/transfer_nft_asset.js
